+++ src/routes/index.js
@@ -5,7 +5,7 @@ import requestRouter from './request.route';
 import bookingRouter from './booking.route';
 import chatRouter from './chat.route';
 import notificationRouter from './notification.route';
-import destinationRouter from './destination.route';
+import statisticsRouter from './statistics.route';
 
 const app = express.Router();
 
@@ -15,6 +15,6 @@ app.use('/requests', requestRouter);
 app.use('/bookings', bookingRouter);
 app.use('/chats', chatRouter);
 app.use('/notifications', notificationRouter);
-app.use('/destinations', destinationRouter);
+app.use('/statistics', statisticsRouter);
 
 export default app;
